Move intervalDescription into the time-diff module
