+++ server/src/pages/shot/view.js
@@ -5,7 +5,7 @@ const ReactDOM = require("react-dom");
 const { Footer } = require("../../footer-view");
 const sendEvent = require("../../browser-send-event.js");
 const { ShareButtons } = require("./share-buttons");
-const { TimeDiff } = require("./time-diff");
+const { TimeDiff, intervalDescription } = require("./time-diff");
 const reactruntime = require("../../reactruntime");
 
 
@@ -560,53 +560,6 @@ class ExpireWidget extends React.Component {
     this.props.onSaveExpire(value);
     this.setState({isChangingExpire: false});
   }
-}
-
-function intervalDescription(ms) {
-  let parts = [];
-  let second = 1000;
-  let minute = second*60;
-  let hour = minute*60;
-  let day = hour*24;
-  if (ms > day) {
-    let days = Math.floor(ms/day);
-    if (days === 1) {
-      parts.push("1 day");
-    } else {
-      parts.push(`${days} days`);
-    }
-    ms = ms % day;
-  }
-  if (ms > hour) {
-    let hours = Math.floor(ms/hour);
-    if (hours === 1) {
-      parts.push("1 hour");
-    } else {
-      parts.push(`{$hours} hours`);
-    }
-    ms = ms % hour;
-  }
-  if (ms > minute) {
-    let minutes = Math.floor(ms/minute);
-    if (minutes === 1) {
-      parts.push("1 minute");
-    } else {
-      parts.push(`${minutes} minutes`);
-    }
-    ms = ms % minute;
-  }
-  if (ms) {
-    let seconds = Math.floor(ms/second);
-    if (seconds === 1) {
-      parts.push("1 second");
-    } else {
-      parts.push(`${seconds} seconds`);
-    }
-  }
-  if (! parts.length) {
-    parts.push("immediately");
-  }
-  return parts.join(" ");
 }
 
 exports.BodyFactory = React.createFactory(Body);
